Refractoring without changing program logic.
* Added classes Helper and Field.
* Ranamed GameServides to Enigine.
* Removed BattleField.
* Methods renamed and moved appropriately.
* Changes to access modifiers and making things static.
