+++ Battle-Field-5/Battle-Field-5 Team Project.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 2013
 VisualStudioVersion = 12.0.31101.0
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BattleField", "BattleField.csproj", "{493704EF-A54B-41A2-B1BF-47857B5337CB}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MineFieldApp", "MineFieldApp.csproj", "{493704EF-A54B-41A2-B1BF-47857B5337CB}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
