+++ Battle-Field-5/EntryPoint.cs
@@ -1,0 +1,10 @@
+﻿namespace MineFieldApp
+{
+    internal static class EntryPoint
+    {
+        internal static void Main()
+        {
+            Engine.Start();
+        }
+    }
+}
